Move authentication components into own folder
