+++ client/src/components/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { NavBar } from './NavBar';
-import { Login } from './Login';
-import { SignUp } from './SignUp';
+import { Login } from './authentication/Login';
+import { SignUp } from './authentication/SignUp';
 import Homepage from './homepage/Homepage';
 
 export default class App extends React.Component {
