+++ client/src/index.js
@@ -3,11 +3,16 @@ import { render } from 'react-dom';
 import { IndexRoute, Router, Route, Link, hashHistory } from 'react-router';
 
 import App from './components/App';
-import SignUp from './components/SignUp';
 import Homepage from './components/homepage/Homepage';
-import LogIn from './components/Login';
+
 import Leaderboard from './components/Leaderboard';
-import {Profile} from './components/Profile';
+import Profile from './components/Profile';
+
+// Authentication
+import SignUp from './components/authentication/SignUp';
+import LogIn from './components/authentication/Login';
+
+// Games
 import GameScramble from './components/games/Scramble/GameScramble';
 import GameMemory from './components/games/GameMemory/GameMemory';
 import TypingSpeedmaster from './components/games/TypingSpeedmaster/TypingSpeedmaster';
